requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.0
langchain==0.3.0
langchain-pinecone==0.2.0
langchain-community==0.3.0
langchain-huggingface==0.1.0
langchain-google-genai==2.0.0
google-generativeai==0.8.0
pinecone-client==5.0.0
transformers==4.45.0
sentence-transformers==3.1.0
